# swagger/swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "local"
    }
  ],
  "info": {
    "title": "Open360 Backend API",
    "version": "1.0.0",
    "description": "SDK-first modular API for tracking, shipments, returns, accounts, and tenancy.",
    "license": {
      "name": "UNLICENSED"
    }
  },
  "components": {
    "schemas": {
...
      "AccountSummary": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "backgroundImage": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "customerGroupId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "billingCode": {
            "type": "number"
          },
          "isMfaEnabled": {
            "type": "boolean"
          },
          "isSsoEnabled": {
            "type": "boolean"
          },
          "standing": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "displayName",
          "status",
          "tenantId",
          "isMfaEnabled",
          "isSsoEnabled",
          "type",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "AccountListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AccountSummary"
            },
            "type": "array"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "type": "object"
      },
...
      "UpdateAccountRequest": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": "string"
          },
          "icon": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "backgroundImage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "about": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "logo": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "status": {
            "oneOf": [
              {
                "const": "ACTIVE"
              },
              {
                "const": "ARCHIVED"
              },
              {
                "const": "INACTIVE"
              },
              {
                "const": "PENDING"
              }
            ]
          },
          "isMfaEnabled": {
            "type": "boolean"
          },
          "isSsoEnabled": {
            "type": "boolean"
          },
          "standing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "const": "HEALTHY"
              },
              {
                "const": "HYPERCARE"
              },
              {
                "const": "NEW"
              },
              {
                "const": "STANDARD"
              }
            ]
          },
          "grade": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "const": "A"
              },
              {
                "const": "B"
              },
              {
                "const": "C"
              },
              {
                "const": "D"
              },
              {
                "const": "X"
              }
            ]
          },
          "type": {
            "oneOf": [
              {
                "const": "B2B"
              },
              {
                "const": "B2C"
              },
              {
                "const": "MANAGED"
              },
              {
                "const": "SAAS"
              }
            ]
          }
        },
        "required": [],
        "type": "object"
      }
    }
  },
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "x-account-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "v1",
          "accounts",
          "list"
        ],
        "x-samchon-controller": "AccountsController"
      }
    },
    "/v1/accounts/by-name/{name}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSummary"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "v1",
          "accounts",
          "by_name",
          "findByName"
        ],
        "x-samchon-controller": "AccountsController"
      }
    },
    "/v1/accounts/{accountId}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "x-account-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSummary"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "v1",
          "accounts",
          "findOne"
        ],
        "x-samchon-controller": "AccountsController"
      },
      "patch": {
        "tags": [],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "x-account-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSummary"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "v1",
          "accounts",
          "update"
        ],
        "x-samchon-controller": "AccountsController"
      }
    }
  },
...
}